cmod_main/scripts/wikidatabots/genewiki/requirements.txt
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
Django==1.6.5
django-celery==3.1.10
django-debug-toolbar==1.2.1
django-extensions==1.3.8
django-grappelli==2.5.3
django-widget-tweaks==1.3
djangorestframework==2.3.14
flake8==2.2.2
gnureadline==6.3.3
gunicorn==19.0.0
html5lib==0.9999999
httplib2==0.9
ipython==2.1.0
isodate==0.5.4
keepalive==0.5
kombu==3.0.21
mccabe==0.2.1
mwclient==0.8.1
mygene==2.1.0
mysqlclient==1.3.7
pep8==1.5.7
pyflakes==0.8.1
pyjade==2.2.0
pyparsing==2.0.7
pytz==2014.4
raven==5.0.0
rdflib==4.2.1
redis==2.10.1
requests==2.10.0
simplejson==3.5.3
six==1.10.0
South==1.0
SPARQLWrapper==1.7.6
sqlparse==0.1.11
Werkzeug==0.9.6
wheel==0.26.0
